dev-requirements.txt
pytest
responses
tox
flake8
black
request
